pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-inference-observability"
version = "0.1.0"
description = "Structured LLM inference benchmarking and JSONL observability hooks"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]
torch = ["torch>=2.1", "transformers>=4.40"]

[project.scripts]
infer-harness = "infer_harness.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
